+++ src/pkg/packager/deploy.go
@@ -319,10 +319,18 @@ func (p *Packager) processComponentFiles(component types.ZarfComponent, pkgLocat
 
 		// If a shasum is specified check it again on deployment as well
 		if file.Shasum != "" {
+			tempFileLocation := fileLocation
+			if file.ArchivePath != "" {
+				fileSource := filepath.Base(file.Source)
+
+				fileLocation, _ = helpers.RenamePathWithFilename(fileLocation, fileSource)
+			}
+
 			spinner.Updatef("Validating SHASUM for %s", file.Target)
 			if shasum, _ := utils.GetCryptoHashFromFile(fileLocation, crypto.SHA256); shasum != file.Shasum {
 				return fmt.Errorf("shasum mismatch for file %s: expected %s, got %s", file.Source, file.Shasum, shasum)
 			}
+			fileLocation = tempFileLocation
 		}
 
 		// Replace temp target directory and home directory
